Sprint 2 Work
First, added createGame button, which then prompts user for text input of the gamename. This info is then displayed as a new game with the user-given info under the game names column. I then added a third column showing when the game was created, as well as a gameCounter that updates every time a new game is created to reflect the number of active games.