Testing feature works now
will switch to a different screen where it shows "test passed" if it can update two cards, discard one of them, play the other, and give a clue to a player. I think play card and discard card is broken so right now it isn't going to work. It shows if invidual tests work in the console.